+++ src/Scene/TexturedBox.h
@@ -12,7 +12,7 @@ public:
 	TexturedBox(const std::vector<GLfloat>& data);
 	~TexturedBox();
 
-	void render(double time, const glm::mat4 &PV) override;
+	void render(float time, const glm::mat4 &PV) override;
 
 	float hOffset = 0.0;
 	float vOffset = 0.0;
